Remove not-selectable SHOULD:populate
https://jira.hl7.org/browse/FHIR-45265

The obligation valueSet (https://build.fhir.org/ig/HL7/fhir-extensions/ValueSet-obligation.html) uses a filter to prevent use/validation of codes identified as 'not-selectable'. We believe not-selectable should be removed and the 'populate' could be modified with the SHOULD modifier in guides that intend to provide best practice recommendations that derivative guides will further constrain for conformance purposes.